pyproject.toml
[project]
name = "ws-ephemeral"
version = "2.2.0"
description = "Request matching ephemeral port"
authors = [
    { name = "Dhruvin Shah", email = "dhruvin3@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "schedule>=1.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "pylint>=3.0.0",
    "mypy>=1.8.0"
]

[tool.pylint.basic]
good-names = ["ws"]
